backend/auth-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
vertica-python==1.3.6
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1